yandex-kafka/producer/requirements.txt
aiohttp==3.8.0
kafka-python==2.0.2
